+++ tests/newtonian/one_dimensional/riemann_problem_3/test.cpp
@@ -71,8 +71,8 @@ namespace {
 void write_output(hdsim1D const& sim, string const& fname)
 {
   ofstream f(fname.c_str());
-  f << sim.GetCell(sim.GetCellNo()/2).Pressure << endl;
-  f << sim.GetCell(sim.GetCellNo()/2).Velocity.x << endl;
+  f << sim.GetCell(static_cast<size_t>(sim.GetCellNo()/2)).Pressure << endl;
+  f << sim.GetCell(static_cast<size_t>(sim.GetCellNo()/2)).Velocity.x << endl;
   f.close();
 }
 }
